requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.6
django-cors-headers==3.13.0
django-nose==1.4.7
djangorestframework==3.14.0
frozenlist==1.3.3
geoip2==4.6.0
idna==3.4
maxminddb==2.2.0
multidict==6.0.4
nose==1.3.7
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
yarl==1.9.1
